Enable all HTTP verbs on the Error controller method
We use UseStatusCodePagesWithReExecute() to generate a 'ProblemDetails' response body for requests that return an empty response with status code signalising an error (>=400). This re-executes the pipeline using the error controller method. However, when the failed request originally used an HTTP verb different from 'GET', this would lead to returning HTTP 415 Method not allowed because the error controller method could only be used with GET requests. This commit fixes this behaviour by modifying the error controller method to accept all HTTP verbs.