Add some suggestions to error messages
It's far from perfect and in some cases there is really nothing to be
said (e.g. when GET for games fails) so at least try to guess what could
be wrong in some cases. However most things should be handled with form
validation and there is no point in informing the user that their
connection is down.

Signed-off-by: František Nečas <frantisek.necas@protonmail.com>