Lock user lookahead after selection
The behaviour of the lookahead is a bit buggy in some browsers on mobile
causing it to query the formatted result again once it's selected
resulting in an error. Editing the chosen user does not make sense (and
resets the model), hence it's better to make the field readonly after
selection anyway and allow resetting it.

Signed-off-by: František Nečas <frantisek.necas@protonmail.com>