Always show the close button of toasts
This makes it more clear that the toast can be dismissed (by tapping
anywhere, not just the button) which should allow the user to faster get
back to using the app when a toast is blocking view of an important
element (e.g. login button blocked by toast mentioning that login has
expired).

Signed-off-by: František Nečas <frantisek.necas@protonmail.com>