Save original start date for recurring states
When planning recurring states, the start date of the meta-state was adjusted to the first actual occurrence even in the saved entity. This caused that the recurring state wouldn't show in the "regular opening hours" immediately but only after the first occurrence - this behaviour was suboptimal from the information point of view. The adjustment has thus been moved to be only used when performing the actual state planning.